requirements.txt
python-dotenv
certifi==2020.6.20
chardet==3.0.4
elasticsearch==7.8.0
idna==2.10
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
tqdm==4.48.0
urllib3==1.25.10
